refactor(cli): deduplicate OutputFormat enum into commands::mod
Move the identical OutputFormat{Text,Json} enum and FromStr impl from
lint.rs and type_coverage.rs into commands/mod.rs. The two copies had
an unintentional punctuation difference in the error message (colon vs
comma); the canonical colon form is kept. Update main.rs to reference
the shared type via commands::OutputFormat.
